Added Multiple Responses For Each Prompt
I added a several possible responses for each statement from the user.
The response is randomly chosen.

I also rearranged the structure of the Android project to get it to
compile.

Finally, I added a Toast to show what was heard from the user. I did
this because the text box at the bottom does not show up on smaller
Android screens.